AAGIPalettes: AAGI Colours and Colour Palettes

Colour palettes based on the official Analytics for the Australian Grains Industry ('AAGI') comms guide and others designed to work harmoniously with the official 'AAGI' colours while being colour vision deficient (CVD) friendly.

See also

Author

Maintainer: Adam H. Sparks adamhsparks@gmail.com (ORCID)

Other contributors:

Curtin University of Technology (Provided support through Adam Sparks's time.) [copyright holder]

Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]