Sequential and Diverging Colour Palettes for AAGI Graphical Outputs

AAGI palettes for use in charts and plots based on official AAGI colours and interpolated to provide more colours for use in maps and other graphical outputs. These palettes are intended to be colour-blind friendly, while still being based on and include official AAGI colours but not all will be official AAGI colours. They are are palettes that are intended to be useful for creating plots that incorporate come official AAGI colours but they may not all be official AAGI colours.

List of available palettes

Diverging palettes

"aagi_BrYl" (11 stops),

"aagi_BuOr" (11 stops),

"aagi_BuYl" (11 stops),

"aagi_GnYl" (11 stops),

"aagi_RdBu" (11 stops),

"aagi_RdTl" (11 stops),

"aagi_RdYl" (11 stops),

"aagi_TlGn" (11 stops),

"aagi_TlYl" (11 stops), and

"aagi_TlGnYl" (11 stops)

Sequential palettes

"aagi_blues" (7 stops),

"aagi_bright_greens" (7 stops),

"aagi_greens" (11 stops),

"aagi_greys" (9 stops),

"aagi_oranges" (9 stops),

"aagi_reds" (9 stops),

"aagi_teals" (9 stops), and

"aagi_yellows" (7 stops)

Usage

aagi_palettes(
  n,
  name = c("aagi_BrYl", "aagi_BuOr", "aagi_BuYl", "aagi_GnYl", "aagi_RdBu", "aagi_RdTl",
    "aagi_RdYl", "aagi_TlGn", "aagi_TlYl", "aagi_TlGnYl", "aagi_blues",
    "aagi_bright_greens", "aagi_greens", "aagi_greys", "aagi_reds", "aagi_teals",
    "aagi_oranges", "aagi_yellows"),
  direction = 1L
)

Arguments

n

Integer Number of different colours in the palette. See details for more.

name

Character A string value with the name of the desired palette.

direction

Integer Sets the order of colours in the scale. If 1, the default order is used. If -1, the order of colours is reversed.

Value

A vector of hexadecimal colour codes

Methods

The colours were taken from the AAGI communications style guide colour palette. In some cases, e.g., aagi_red, #8b0b00 value, did not exist in the official palette and was generated by taking the complimentary colour of the aagi_teal, #00808b value. The darkest and lightest values used in interpolating the colours were selected from the "Tint Colour Variation" list of this website, https://www.colorhexa.com/c5856d, using the AAGI hexadecimal colour code as the input. The interpolated values were all generated using http://www.vis4.net/palettes/, ensuring that all palettes are colourblind safe.

When name is any other palette, the minimum value for n is 3 and maximum will vary from 7 to 9 or 11 stops based on the palette requested.

Examples

# sequential blues with 7 stops
aagi_palettes(n = 7, name = "aagi_blues", direction = 1)
#> [1] "#648fd2" "#81a0da" "#9bb2e1" "#b4c4e8" "#ccd7f0" "#e4eaf7" "#fcfdfe"

# diverging red teal colours with 11 stops
aagi_palettes(n = 11, name = "aagi_RdTl", direction = 1)
#>  [1] "#8b0b00" "#a8432d" "#c36d59" "#da9788" "#eec2b9" "#ffeeed" "#c6e4e7"
#>  [8] "#9ecacf" "#76b1b8" "#4b98a1" "#00808b"